refactor: :recycle: Reorganize project structure into modular files (agent_node.py, graph.py, main.py)
