+++ app/state.py
@@ -1,0 +1,8 @@
+
+from typing import TypedDict
+
+
+class CalculatorState(TypedDict):
+    first_number: int
+    second_number: int
+    result: int
